+++ frontend/src/metabase/querying/filters/components/FilterPanel/FilterPanelPopover/FilterPanelPopover.tsx
@@ -1,7 +1,7 @@
 import { useMemo, useState } from "react";
 
 import { useTranslateContent } from "metabase/i18n/hooks";
-import { PLUGIN_CONTENT_TRANSLATION } from "metabase/plugins";
+import { getTranslatedFilterDisplayName } from "metabase/querying/filters/utils/display";
 import { Popover } from "metabase/ui";
 import * as Lib from "metabase-lib";
 
@@ -26,19 +26,10 @@ export function FilterPanelPopover({
   const [isOpened, setIsOpened] = useState(false);
   const tc = useTranslateContent();
 
-  const translatedFilterName = useMemo(() => {
-    const displayInfo = Lib.displayInfo(query, stageIndex, filter);
-    const parts = Lib.filterParts(query, stageIndex, filter);
-    const columnDisplayName = parts?.column
-      ? Lib.displayInfo(query, stageIndex, parts.column).displayName
-      : undefined;
-
-    return PLUGIN_CONTENT_TRANSLATION.getTranslatedFilterDisplayName(
-      displayInfo.longDisplayName ?? "",
-      tc,
-      columnDisplayName,
-    );
-  }, [query, stageIndex, filter, tc]);
+  const translatedFilterName = useMemo(
+    () => getTranslatedFilterDisplayName(query, stageIndex, filter, tc),
+    [query, stageIndex, filter, tc],
+  );
 
   const handleChange = (newFilter: Lib.Clause | Lib.SegmentMetadata) => {
     onChange(Lib.replaceClause(query, stageIndex, filter, newFilter));
