+++ frontend/src/metabase/querying/segments/components/SegmentFilterEditor/SegmentFilterEditor.tsx
@@ -3,8 +3,8 @@ import { t } from "ttag";
 
 import ErrorBoundary from "metabase/ErrorBoundary";
 import { useTranslateContent } from "metabase/i18n/hooks";
-import { PLUGIN_CONTENT_TRANSLATION } from "metabase/plugins";
 import { FilterPicker } from "metabase/querying/filters/components/FilterPicker";
+import { getTranslatedFilterDisplayName } from "metabase/querying/filters/utils/display";
 import { ClauseStep } from "metabase/querying/notebook/components/ClauseStep";
 import * as Lib from "metabase-lib";
 
@@ -25,19 +25,8 @@ export function SegmentFilterEditor({
   const filters = useMemo(() => Lib.filters(query, STAGE_INDEX), [query]);
 
   const renderFilterName = useCallback(
-    (filter: Lib.FilterClause) => {
-      const displayInfo = Lib.displayInfo(query, STAGE_INDEX, filter);
-      const parts = Lib.filterParts(query, STAGE_INDEX, filter);
-      const columnDisplayName = parts?.column
-        ? Lib.displayInfo(query, STAGE_INDEX, parts.column).displayName
-        : undefined;
-
-      return PLUGIN_CONTENT_TRANSLATION.getTranslatedFilterDisplayName(
-        displayInfo.longDisplayName ?? "",
-        tc,
-        columnDisplayName,
-      );
-    },
+    (filter: Lib.FilterClause) =>
+      getTranslatedFilterDisplayName(query, STAGE_INDEX, filter, tc),
     [query, tc],
   );
